// src/core/prompts/system-prompt/types.ts
export const TASK_PROGRESS_PARAMETER = {
    name: "task_progress",
    required: false,
    instruction: `A checklist showing task progress after this tool use is completed. (See 'Updating Task Progress' section for more details)`,
    usage: "Checklist here (optional)",
    dependencies: [ClineDefaultTool.TODO],
}